+++ Src/Analyzers/TrailingWhitespaceAnalyzer.cs
@@ -8,9 +8,9 @@ namespace CodeHeroes.CodeAnalysis.Style
     using Microsoft.CodeAnalysis.Text;
 
     [DiagnosticAnalyzer(LanguageNames.CSharp)]
-    public sealed class TrailingWhitespaceDiagnosticAnalyzer : DiagnosticAnalyzer
+    public sealed class TrailingWhitespaceAnalyzer : DiagnosticAnalyzer
     {
-        public const string DiagnosticId = "TrailingWhitespace";
+        public const string DiagnosticId = "CH0001";
 
         private static DiagnosticDescriptor Rule = new DiagnosticDescriptor(
             DiagnosticId,