+++ release-notes.md
@@ -1,3 +1,10 @@
+## New in 0.0.6 (Released November 3rd, 2017)
+* feature: using directives within namespace analyzer
+* feature: using directives sorted correctly analyzer
+* feature: single namespace per file analyzer
+* feature: single top-level type per file analyzer (delegates excepted)
+* feature: reserve underscore identifers for discard semantics analyzer
+
 ## New in 0.0.5 (Released October 31st, 2017)
 * fix: ignore generated code
 * chore: rename to CodeHeroes.CodeAnalysis.Style
@@ -10,7 +17,7 @@
 * fix: allow analyzers to be installed in non-SDK-style csproj
 
 ## New in 0.0.2 (Released October 26th, 2017)
-* feature: trailing whitespace visualizer
+* feature: trailing whitespace analyzer
 
 ## New in 0.0.1 (Released October 23rd, 2017)
 * chore: initial test release